refactor(credentials): retire old per-surface credential routes
The unified Credentials page was importing its dialog components from the old
per-surface route files, so those routes (and their RBAC rows) had to stay
alive — two live URLs for the same feature. Finish the consolidation:

- Extract the OAuth client Create/Edit dialogs into co-located non-route
  modules (credentials/_parts/{mcp,llm}-oauth-client-dialogs.tsx), and unify
  the post-create/rotate 'secret revealed' dialog into one component
  parameterized by scope. This also fixes a latent bug where an LLM
  authorization-code client showed the 'mcp' scope hint instead of 'llm:proxy'.
- Move the virtual-keys page (and its owner-select-field) under
  credentials/virtual-keys, so it is a real page rather than a re-export.
- Delete the old /llm/credentials and /mcp/credentials route trees and their
  now-dead requiredPagePermissionsMap entries.

Net -800 lines; type-check, lint, knip, and tests green.
