+++ platform/frontend/src/app/credentials/_parts/mcp-oauth-client-dialogs.tsx
@@ -4,24 +4,14 @@ import type {
   archestraApiTypes,
   ResourceVisibilityScope,
 } from "@archestra/shared";
-import type { ColumnDef } from "@tanstack/react-table";
-import { KeyRound, Pencil, Plus, RefreshCw, Trash2 } from "lucide-react";
-import { useEffect, useMemo, useState } from "react";
+import { useEffect, useState } from "react";
 import {
   AgentSelector,
   type AgentSelectorAgent,
 } from "@/components/agent-selector";
-import { CopyableCode } from "@/components/copyable-code";
-import { DeleteConfirmDialog } from "@/components/delete-confirm-dialog";
 import { FormDialog } from "@/components/form-dialog";
 import { OauthClientVisibilityField } from "@/components/oauth-client-visibility-field";
-import { QueryLoadError } from "@/components/query-load-error";
-import { ResourceVisibilityBadge } from "@/components/resource-visibility-badge";
-import { SearchInput } from "@/components/search-input";
-import { TableRowActions } from "@/components/table-row-actions";
-import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
-import { DataTable } from "@/components/ui/data-table";
 import {
   DialogBody,
   DialogForm,
@@ -31,27 +21,10 @@ import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
 import { Textarea } from "@/components/ui/textarea";
-import { useProfiles } from "@/lib/agent.query";
-import { useSession } from "@/lib/auth/auth.query";
-import { useDataTableQueryParams } from "@/lib/hooks/use-data-table-query-params";
-import {
-  useCreateMcpOauthClient,
-  useDeleteMcpOauthClient,
-  useMcpOauthClients,
-  useRotateMcpOauthClientSecret,
-  useUpdateMcpOauthClient,
-} from "@/lib/mcp-oauth-clients.query";
-import { formatRelativeTimeFromNow } from "@/lib/utils/date-time";
-import { useSetCredentialsAction } from "../layout";
 
 export type McpOauthClient =
   archestraApiTypes.GetMcpOauthClientsResponses["200"][number];
 type GrantType = McpOauthClient["grantType"];
-type CreatedCredentials = {
-  clientId: string;
-  clientSecret: string;
-  grantType: GrantType;
-};
 
 const GRANT_TYPE_OPTIONS: {
   value: GrantType;
@@ -71,293 +44,6 @@ const GRANT_TYPE_OPTIONS: {
       "A pre-registered app obtains user-scoped tokens, so gateway tools resolve each user's own identity and connections.",
   },
 ];
-
-const GRANT_TYPE_LABEL: Record<GrantType, string> = {
-  client_credentials: "Application",
-  authorization_code: "User-delegated",
-};
-
-function parseRedirectUris(text: string): string[] {
-  return text
-    .split("\n")
-    .map((line) => line.trim())
-    .filter(Boolean);
-}
-
-export default function OAuthClientsPage() {
-  const { searchParams, updateQueryParams } = useDataTableQueryParams();
-  const search = searchParams.get("search") || "";
-
-  const {
-    data: oauthClients = [],
-    isPending,
-    isLoadingError: isOauthClientsLoadError,
-    refetch: refetchOauthClients,
-  } = useMcpOauthClients({
-    search: search || undefined,
-  });
-  // MCP gateways and A2A agents are both reached by id through the same token
-  // validation, so an OAuth client can be scoped to either.
-  const { data: gateways = [] } = useProfiles({
-    filters: { agentTypes: ["mcp_gateway", "agent"] },
-  });
-  const createMutation = useCreateMcpOauthClient();
-  const updateMutation = useUpdateMcpOauthClient();
-  const rotateMutation = useRotateMcpOauthClientSecret();
-  const deleteMutation = useDeleteMcpOauthClient();
-  const { data: session } = useSession();
-  const currentUserId = session?.user?.id;
-
-  const [isCreateDialogOpen, setIsCreateDialogOpen] = useState(false);
-  const [createdCredentials, setCreatedCredentials] =
-    useState<CreatedCredentials | null>(null);
-  const [deletingOAuthClient, setDeletingOAuthClient] =
-    useState<McpOauthClient | null>(null);
-  const [editingOAuthClient, setEditingOAuthClient] =
-    useState<McpOauthClient | null>(null);
-  const [rotatedCredentials, setRotatedCredentials] =
-    useState<CreatedCredentials | null>(null);
-  const [rotatingOAuthClient, setRotatingOAuthClient] =
-    useState<McpOauthClient | null>(null);
-
-  const setCredentialsAction = useSetCredentialsAction();
-  useEffect(() => {
-    setCredentialsAction(
-      <Button onClick={() => setIsCreateDialogOpen(true)}>
-        <Plus className="h-4 w-4" />
-        Create OAuth Client
-      </Button>,
-    );
-    return () => setCredentialsAction(null);
-  }, [setCredentialsAction]);
-
-  const columns: ColumnDef<McpOauthClient>[] = useMemo(
-    () => [
-      {
-        accessorKey: "name",
-        header: "Name",
-        cell: ({ row }) => (
-          <div className="font-medium">{row.original.name}</div>
-        ),
-      },
-      {
-        accessorKey: "clientId",
-        header: "Client ID",
-        cell: ({ row }) => (
-          <code className="text-xs text-muted-foreground">
-            {row.original.clientId}
-          </code>
-        ),
-      },
-      {
-        id: "grantType",
-        header: "Type",
-        cell: ({ row }) => (
-          <Badge variant="secondary">
-            {GRANT_TYPE_LABEL[row.original.grantType]}
-          </Badge>
-        ),
-      },
-      {
-        id: "scope",
-        header: "Gateways / Redirects",
-        cell: ({ row }) => (
-          <span className="text-sm text-muted-foreground">
-            {row.original.grantType === "authorization_code"
-              ? `${row.original.redirectUris.length} redirect URI(s)`
-              : `${row.original.allowedGatewayIds.length} gateway(s)`}
-          </span>
-        ),
-      },
-      {
-        id: "visibility",
-        header: "Accessible to",
-        cell: ({ row }) => (
-          <ResourceVisibilityBadge
-            scope={row.original.scope}
-            teams={row.original.teams}
-            authorId={row.original.authorId}
-            authorName={row.original.authorName}
-            currentUserId={currentUserId}
-          />
-        ),
-      },
-      {
-        accessorKey: "createdAt",
-        header: "Created",
-        cell: ({ row }) => (
-          <span className="text-sm text-muted-foreground">
-            {formatRelativeTimeFromNow(row.original.createdAt)}
-          </span>
-        ),
-      },
-      {
-        id: "actions",
-        header: "Actions",
-        cell: ({ row }) => (
-          <TableRowActions
-            actions={[
-              {
-                icon: <Pencil className="h-4 w-4" />,
-                label: "Edit",
-                onClick: () => setEditingOAuthClient(row.original),
-              },
-              {
-                icon: <RefreshCw className="h-4 w-4" />,
-                label: "Rotate secret",
-                onClick: () => setRotatingOAuthClient(row.original),
-              },
-              {
-                icon: <Trash2 className="h-4 w-4" />,
-                label: "Delete",
-                variant: "destructive",
-                onClick: () => setDeletingOAuthClient(row.original),
-              },
-            ]}
-          />
-        ),
-      },
-    ],
-    [currentUserId],
-  );
-
-  return (
-    <>
-      <div className="mb-4 flex flex-wrap gap-4">
-        <SearchInput
-          objectNamePlural="OAuth clients"
-          searchFields={["name"]}
-          paramName="search"
-        />
-      </div>
-
-      {isOauthClientsLoadError ? (
-        <QueryLoadError
-          title="Couldn't load your OAuth clients"
-          onRetry={() => refetchOauthClients()}
-        />
-      ) : (
-        <DataTable
-          columns={columns}
-          data={oauthClients}
-          isLoading={isPending}
-          emptyMessage="No OAuth clients registered. Create one for an application that calls MCP gateways or agents."
-          hasActiveFilters={Boolean(search)}
-          filteredEmptyMessage="No OAuth clients match your filters. Try adjusting your search."
-          onClearFilters={() =>
-            updateQueryParams({
-              search: null,
-              page: "1",
-            })
-          }
-        />
-      )}
-
-      <CreateOAuthClientDialog
-        open={isCreateDialogOpen}
-        onOpenChange={setIsCreateDialogOpen}
-        gateways={gateways}
-        onSubmit={async (values) => {
-          const result = await createMutation.mutateAsync(values);
-          if (result) {
-            setCreatedCredentials({
-              clientId: result.clientId,
-              clientSecret: result.clientSecret,
-              grantType: result.grantType,
-            });
-            setIsCreateDialogOpen(false);
-          }
-        }}
-        isSubmitting={createMutation.isPending}
-      />
-
-      <EditOAuthClientDialog
-        oauthClient={editingOAuthClient}
-        onOpenChange={(open) => {
-          if (!open) setEditingOAuthClient(null);
-        }}
-        gateways={gateways}
-        onSubmit={async (id, values) => {
-          const result = await updateMutation.mutateAsync({
-            id,
-            body: values,
-          });
-          if (result) {
-            setEditingOAuthClient(null);
-          }
-        }}
-        isSubmitting={updateMutation.isPending}
-      />
-
-      <CredentialsDialog
-        open={!!createdCredentials}
-        onOpenChange={(open) => {
-          if (!open) setCreatedCredentials(null);
-        }}
-        title="OAuth Client Created"
-        credentials={createdCredentials}
-      />
-
-      <CredentialsDialog
-        open={!!rotatedCredentials}
-        onOpenChange={(open) => {
-          if (!open) setRotatedCredentials(null);
-        }}
-        title="Secret Rotated"
-        credentials={rotatedCredentials}
-      />
-
-      <DeleteConfirmDialog
-        open={!!rotatingOAuthClient}
-        onOpenChange={(open) => {
-          if (!open) setRotatingOAuthClient(null);
-        }}
-        title="Rotate OAuth client secret"
-        description={
-          rotatingOAuthClient
-            ? `Rotate the secret for ${rotatingOAuthClient.name}? Existing integrations using the current secret will not be able to request new access tokens.`
-            : ""
-        }
-        onConfirm={async () => {
-          if (!rotatingOAuthClient) return;
-          const result = await rotateMutation.mutateAsync({
-            id: rotatingOAuthClient.id,
-          });
-          if (result) {
-            setRotatedCredentials({
-              clientId: result.clientId,
-              clientSecret: result.clientSecret,
-              grantType: result.grantType,
-            });
-          }
-          setRotatingOAuthClient(null);
-        }}
-        isPending={rotateMutation.isPending}
-        confirmLabel="Rotate secret"
-        pendingLabel="Rotating..."
-      />
-
-      <DeleteConfirmDialog
-        open={!!deletingOAuthClient}
-        onOpenChange={(open) => {
-          if (!open) setDeletingOAuthClient(null);
-        }}
-        title="Delete OAuth client"
-        description={
-          deletingOAuthClient
-            ? `Delete ${deletingOAuthClient.name}? Existing access tokens will stop working when they expire, and new tokens cannot be issued.`
-            : ""
-        }
-        onConfirm={async () => {
-          if (!deletingOAuthClient) return;
-          await deleteMutation.mutateAsync({ id: deletingOAuthClient.id });
-          setDeletingOAuthClient(null);
-        }}
-        isPending={deleteMutation.isPending}
-      />
-    </>
-  );
-}
 
 export function CreateOAuthClientDialog({
   open,
@@ -617,6 +303,17 @@ export function EditOAuthClientDialog({
   );
 }
 
+// ===
+// Internal helpers
+// ===
+
+function parseRedirectUris(text: string): string[] {
+  return text
+    .split("\n")
+    .map((line) => line.trim())
+    .filter(Boolean);
+}
+
 function GrantTypeField({
   value,
   onChange,
@@ -712,72 +409,5 @@ function GatewayGrantField({
         permissions).
       </p>
     </div>
-  );
-}
-
-export function CredentialsDialog({
-  open,
-  onOpenChange,
-  title,
-  credentials,
-}: {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  title: string;
-  credentials: CreatedCredentials | null;
-}) {
-  const endpoint = (path: string) =>
-    typeof window === "undefined"
-      ? path
-      : new URL(path, window.location.origin).toString();
-  const isAuthorizationCode = credentials?.grantType === "authorization_code";
-
-  return (
-    <FormDialog
-      open={open}
-      onOpenChange={onOpenChange}
-      title={title}
-      description="Copy the client secret now. It will not be shown again."
-    >
-      <DialogBody className="space-y-4">
-        {credentials && (
-          <>
-            <div className="space-y-2">
-              <Label>Client ID</Label>
-              <CopyableCode value={credentials.clientId} />
-            </div>
-            <div className="space-y-2">
-              <Label>Client Secret</Label>
-              <CopyableCode value={credentials.clientSecret} />
-            </div>
-            {isAuthorizationCode && (
-              <div className="rounded-md border bg-muted/40 p-3 text-sm">
-                <div className="mb-2 flex items-center gap-2 font-medium">
-                  <KeyRound className="h-4 w-4" />
-                  Authorization endpoint
-                </div>
-                <CopyableCode value={endpoint("/api/auth/oauth2/authorize")} />
-                <p className="mt-2 text-muted-foreground">
-                  Use the authorization code flow with PKCE and the{" "}
-                  <code>mcp</code> scope.
-                </p>
-              </div>
-            )}
-            <div className="rounded-md border bg-muted/40 p-3 text-sm">
-              <div className="mb-2 flex items-center gap-2 font-medium">
-                <KeyRound className="h-4 w-4" />
-                Token endpoint
-              </div>
-              <CopyableCode value={endpoint("/api/auth/oauth2/token")} />
-            </div>
-          </>
-        )}
-      </DialogBody>
-      <DialogStickyFooter>
-        <Button type="button" onClick={() => onOpenChange(false)}>
-          Done
-        </Button>
-      </DialogStickyFooter>
-    </FormDialog>
   );
 }
